requirements.txt
fastapi==0.95.2
uvicorn==0.21.1
gunicorn==20.1.0
python-telegram-bot==20.0
pydantic==1.10.7
python-dotenv==0.21.0
httpx==0.23.3
